# Improve OvalBuilder Clone Performance